Rename template config and move transformations to code snippets
