+++ docs/Extensibility_UG/matcher_pass.md
@@ -3,9 +3,9 @@
 `ov::pass::MatcherPass` is used for pattern-based transformations.
 
 Template for MatcherPass transformation class
-@snippet src/transformations/template_pattern_transformation.hpp graph_rewrite:template_transformation_hpp
+@snippet template_pattern_transformation.hpp graph_rewrite:template_transformation_hpp
 
-@snippet src/transformations/template_pattern_transformation.cpp graph_rewrite:template_transformation_cpp
+@snippet template_pattern_transformation.cpp graph_rewrite:template_transformation_cpp
 
 To use `ov::pass::MatcherPass`, you need to complete these steps:
 1. Create a pattern
