+++ docs/snippets/template_model_transformation.hpp
