+++ docs/snippets/template_pattern_transformation.hpp
