+++ src/plugins/template/src/compiled_model.hpp
@@ -4,11 +4,11 @@
 
 #pragma once
 
+#include "config.hpp"
 #include "openvino/runtime/icompiled_model.hpp"
 #include "openvino/runtime/iinfer_request.hpp"
 #include "openvino/runtime/isync_infer_request.hpp"
 #include "openvino/runtime/tensor.hpp"
-#include "template_config.hpp"
 
 namespace ov {
 namespace template_plugin {
